feat: initialize mdx
