+++ package.json
@@ -15,6 +15,8 @@
     "@fortawesome/react-fontawesome": "^0.1.18",
     "@headlessui/react": "^1.5.0",
     "@heroicons/react": "^1.0.6",
+    "@mdx-js/loader": "^2.1.0",
+    "@next/mdx": "^12.1.1",
     "axios": "^0.26.1",
     "chrome-aws-lambda": "^10.1.0",
     "next": "^12.1.0",
@@ -24,11 +26,13 @@
     "react-markdown": "^8.0.1",
     "react-syntax-highlighter": "^15.5.0",
     "react-toastify": "^8.2.0",
+    "rehype-highlight": "^5.0.2",
     "rehype-raw": "^6.1.1",
     "rehype-slug": "^5.0.1",
     "swr": "^1.2.2"
   },
   "devDependencies": {
+    "@tailwindcss/typography": "^0.5.2",
     "@types/node": "17.0.23",
     "@types/react": "17.0.43",
     "@types/react-syntax-highlighter": "^13.5.2",
