+++ src/modules/blog/components/Comment/CommentSection.tsx
@@ -2,7 +2,7 @@ import Modal from '@/common/components/Modal'
 import SpinnerLoading from '@/common/components/SpinnerLoading'
 import ModalAction from '@/common/types/ModalAction'
 import React, { FC, useState } from 'react'
-import { Post } from '../ViewPostPage'
+import { Post } from '../Layout/PostLayout'
 import CommentCard from './CommentCard'
 import NewCommentSection from './NewCommentSection'
 import useComments from './useComments'
