--- src/pages/blog/posts/[slug].tsx
@@ -1,5 +1,0 @@
-export {
-  default,
-  getStaticPaths,
-  getStaticProps,
-} from '@/modules/blog/components/ViewPostPage'
